+++ src/app/(main)/settings/llm/index.tsx
@@ -1,13 +1,7 @@
 'use client';
 
-import Link from 'next/link';
-import { memo } from 'react';
-import { Trans, useTranslation } from 'react-i18next';
+import { Flexbox } from 'react-layout-kit';
 
-import PageTitle from '@/components/PageTitle';
-import { MORE_MODEL_PROVIDER_REQUEST_URL } from '@/const/url';
-
-import Footer from '../features/Footer';
 import Anthropic from './Anthropic';
 import Azure from './Azure';
 import Bedrock from './Bedrock';
@@ -23,13 +17,11 @@ import Perplexity from './Perplexity';
 import TogetherAI from './TogetherAI';
 import ZeroOne from './ZeroOne';
 import Zhipu from './Zhipu';
+import Footer from './components/Footer';
 
-export default memo(() => {
-  const { t } = useTranslation('setting');
-
+const Page = () => {
   return (
-    <>
-      <PageTitle title={t('tab.llm')} />
+    <Flexbox gap={24}>
       <OpenAI />
       <Ollama />
       <Azure />
@@ -45,15 +37,11 @@ export default memo(() => {
       <Moonshot />
       <Zhipu />
       <ZeroOne />
-      <Footer>
-        <Trans i18nKey="llm.waitingForMore" ns={'setting'}>
-          更多模型正在
-          <Link aria-label={'todo'} href={MORE_MODEL_PROVIDER_REQUEST_URL} target="_blank">
-            计划接入
-          </Link>
-          中 ，敬请期待 ✨
-        </Trans>
-      </Footer>
-    </>
+      <Footer />
+    </Flexbox>
   );
-});
+};
+
+Page.displayName = 'LlmSetting';
+
+export default Page;