+++ src/features/AvatarWithUpload/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Upload } from 'antd';
 import { createStyles } from 'antd-style';
 import NextImage from 'next/image';